+++ cxbin/plugin/plugincxbin.h
@@ -22,8 +22,8 @@ namespace cxbin
 		virtual ~CXBinLoader();
 
 		virtual std::string expectExtension() override;
-		virtual bool tryLoad(FILE* f, unsigned long long fileSize) override;
-		virtual bool load(FILE* f, unsigned fileSize, std::vector<trimesh::TriMesh*>& out, ccglobal::Tracer* tracer) override;
+		virtual bool tryLoad(FILE* f, size_t fileSize) override;
+		virtual bool load(FILE* f, size_t fileSize, std::vector<trimesh::TriMesh*>& out, ccglobal::Tracer* tracer) override;
 
 	};
 }
